+++ src/TestSuiteStorage/Action/PortalNodeConfigurationTestContract.php
@@ -59,9 +59,8 @@ abstract class PortalNodeConfigurationTestContract extends TestCase
             new PortalNodeCreatePayload(PortalC::class()),
         ]);
         $createResults = $createAction->create($createPayloads);
-        $portalNodeKeys = new PortalNodeKeyCollection(\iterable_map(
-            $createResults,
-            static fn (PortalNodeCreateResult $r): PortalNodeKeyInterface => $r->getPortalNodeKey()
+        $portalNodeKeys = new PortalNodeKeyCollection($createResults->map(
+            static fn (PortalNodeCreateResult $result): PortalNodeKeyInterface => $result->getPortalNodeKey()
         ));
         $getCriteria = new PortalNodeConfigurationGetCriteria($portalNodeKeys);
 
@@ -70,7 +69,7 @@ abstract class PortalNodeConfigurationTestContract extends TestCase
         foreach ($portalNodeKeys as $step => $portalNodeKey) {
             static::assertCount($step, \iterable_to_array(\iterable_filter(
                 $getAction->get($getCriteria),
-                static fn (PortalNodeConfigurationGetResult $g): bool => $g->getValue() !== []
+                static fn (PortalNodeConfigurationGetResult $result): bool => $result->getValue() !== []
             )));
 
             $setAction->set(new PortalNodeConfigurationSetPayloads([
@@ -79,7 +78,7 @@ abstract class PortalNodeConfigurationTestContract extends TestCase
 
             static::assertCount($step + 1, \iterable_to_array(\iterable_filter(
                 $getAction->get($getCriteria),
-                static fn (PortalNodeConfigurationGetResult $g): bool => $g->getValue() !== []
+                static fn (PortalNodeConfigurationGetResult $result): bool => $result->getValue() !== []
             )));
 
             /** @var PortalNodeConfigurationGetResult[] $getResults */
@@ -88,13 +87,16 @@ abstract class PortalNodeConfigurationTestContract extends TestCase
             static::assertTrue($getResults[0]->getPortalNodeKey()->equals($portalNodeKey));
 
             $readConfiguration = $getResults[0]->getValue();
+
+            static::assertArrayHasKey('object', $readConfiguration);
+
             \ksort($readConfiguration['object']);
             static::assertSame($testPayload, $readConfiguration);
         }
 
         static::assertCount($portalNodeKeys->count(), \iterable_to_array(\iterable_filter(
             $getAction->get($getCriteria),
-            static fn (PortalNodeConfigurationGetResult $g): bool => $g->getValue() !== []
+            static fn (PortalNodeConfigurationGetResult $result): bool => $result->getValue() !== []
         )));
 
         $setPayload = new PortalNodeConfigurationSetPayloads();
@@ -107,7 +109,7 @@ abstract class PortalNodeConfigurationTestContract extends TestCase
 
         static::assertEmpty(\iterable_to_array(\iterable_filter(
             $getAction->get($getCriteria),
-            static fn (PortalNodeConfigurationGetResult $g): bool => $g->getValue() !== []
+            static fn (PortalNodeConfigurationGetResult $result): bool => $result->getValue() !== []
         )));
 
         $deleteAction->delete(new PortalNodeDeleteCriteria($portalNodeKeys));