+++ src/TestSuiteStorage/Action/RouteTestContract.php
@@ -45,6 +45,9 @@ use Heptacom\HeptaConnect\TestSuite\Storage\TestCase;
 
 /**
  * Test pre-implementation to test route related storage actions. Some other storage actions e.g. PortalNodeCreate are needed to set up test scenarios.
+ *
+ * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+ * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
  */
 abstract class RouteTestContract extends TestCase
 {