+++ src/UiAdminBase/Action/PortalNode/PortalNodeEntityList/PortalNodeEntityListResult.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeEntityList;
+
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\EntityListResultContract;
+
+final class PortalNodeEntityListResult extends EntityListResultContract
+{
+}