+++ src/TestSuiteStorage/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add test scenario in `\Heptacom\HeptaConnect\TestSuite\Storage\Action\IdentityMappingTestContract::testReflectFromPortalNodeAToBWhereNoMappingsAreInTheStorage` to test the mapping functionality of identity reflection action implementations
+- Add admin UI action fixture class `\Heptacom\HeptaConnect\TestSuite\Storage\Fixture\AdminUiAction\StorageTestUiAction` to provide a basic admin UI action
+- Add `\Heptacom\HeptaConnect\TestSuite\Storage\Action\UiAuditTrailTestContract` to test basic life cycle of logging a UI action
 - Add tests in `\Heptacom\HeptaConnect\TestSuite\Storage\Action\RouteTestContract` to test `\Heptacom\HeptaConnect\Storage\Base\Contract\Action\Route\RouteOverviewActionInterface` usage for different sorting and pagination behaviours
 
 ### Changed